agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
# livekit-plugins-silero # Temporarily disabled due to compatibility issues
python-dotenv
requests
